requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
oauthlib==3.2.0
python-dotenv==0.20.0
requests==2.27.1
requests-oauthlib==1.3.1
tweepy==4.8.0
urllib3==1.26.9
